fix review points
